requirements.txt
# Base
psycopg2-binary==2.8.4
celery==4.3.0
redis==3.4.1
PyYAML==5.3.1
coverage==5.0.4

# Web application
Django==2.2.11
djangorestframework==3.11.0
drf-yasg==1.17.1

# Algorithm Service
flask==1.1.1
Flask-SQLAlchemy==2.4.1
numpy==1.12.1
pandas==0.19.2
scikit-learn==0.18.1
scipy==0.19.0
xlrd==1.1.0
